+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue where deleted notes still appeared in the Profile’s Notes view.
 - Sorted the featured profiles in the Discover tab.
 - Switched from Reportinator to Tagr bot for content labeling.
+- Discover tab now features new accounts.
 
 ## [0.1.14] - 2024-05-22Z